New HMM recipe working.

+++ recipes/timit_v2/steps/train_hmm2.sh
@@ -27,7 +27,7 @@ if [ ! -d $mdl_dir ];then
     cp $hmm_emission_conf $mdl_dir
 fi
 
-if [ ! -f $mdl_dir/init.mdl ]; then
+if [ ! -f $mdl_dir/0.mdl ]; then
     echo "Building the HMM model..."
     # Build the decoding graph.
     python utils/prepare-decode-graph.py \
@@ -52,13 +52,13 @@ if [ ! -f $mdl_dir/init.mdl ]; then
         $mdl_dir/emissions.mdl \
         $mdl_dir/0.mdl || exit 1
 else
-    echo "Using previous created HMM: $mdl_dir/init.mdl"
+    echo "Using previous created HMM: $mdl_dir/0.mdl"
 fi
 
 
 # Prepare the alignments.
 if [ ! -f $mdl_dir/ali_graphs.npz ]; then
-    echo "Preparing the alignment graph..."
+    echo "Preparing alignment graphs..."
 
     tmpdir=$(mktemp -d /tmp/beer.XXXX);
     trap 'rm -rf "$tmpdir"' EXIT
@@ -70,8 +70,9 @@ if [ ! -f $mdl_dir/ali_graphs.npz ]; then
     find $tmpdir -name '*npy' \
         | zip -j -@ $mdl_dir/ali_graphs.npz > /dev/null || exit 1
 else
-    echo "Alignment already prepared: $mdl_dir/ali.npz"
+    echo "Alignment graphs already prepared: $mdl_dir/ali_graphs.npz"
 fi
+
 
 if [ ! -f $mdl_dir/final.mdl ];then
     echo "Training HMM-GMM model"
@@ -84,67 +85,56 @@ if [ ! -f $mdl_dir/final.mdl ];then
     mkdir -p $mdl_dir/split
     cp $data_train_dir/uttids $mdl_dir/split/uttids
     pushd $mdl_dir/split > /dev/null
-    split --numeric-suffixes=1 -n l/$hmm_ali_njobs ./uttids
+    split --numeric-suffixes=1 -n l/$hmm_align_njobs ./uttids
     popd > /dev/null
 
-    # Retrieve the last model (ordered by number).
+    # Retrieve the last model.
     mdl=$(find $mdl_dir -name "[0-9]*mdl" -exec basename {} \; | \
         sort -t '.' -k 1 -g | tail -1)
-    align_no="${mdl%.*}"
+    epoch="${mdl%.*}"
 
-    while [ $align_no -le $hmm_n_align ]; do
-        tmpdir=$(mktemp -d $mdl_dir/tmp.XXXX);
-        trap 'rm -rf "$tmpdir"' EXIT
+    while [ $((++epoch)) -le $hmm_train_epochs ]; do
+        echo "Iteration: $epoch"
 
-        echo "Iteration: $align_no"
+        if echo $hmm_align_epochs | grep -w $epoch >/dev/null; then
+            echo "Aligning data"
 
-        if [ ! -f $mdl_dir/alis.$align_no.npz ]; then
-            echo "Aligning the features to the transcription"
+            tmpdir=$(mktemp -d $mdl_dir/tmp.XXXX);
+            trap 'rm -rf "$tmpdir"' EXIT
+
             cmd="python utils/hmm-align.py \
                 --utt-graphs $mdl_dir/ali_graphs.npz \
-                $mdl_dir/$mdl \
-                $data_train_dir/feats.npz \
-                $tmpdir"
-            qsub \
-                -N "beer-align" \
-                -cwd \
-                -j y \
-                -o $mdl_dir/log/align.$align_no.out.'$TASK_ID' \
-                -t 1-$hmm_ali_njobs \
-                -sync y \
-                $fea_sge_opts \
+                $mdl_dir/$mdl  $data_train_dir/feats.npz  $tmpdir"
+            qsub  -N "beer-align"  -cwd  -j y -sync y $hmm_align_sge_opts \
+                -o $mdl_dir/log/align.${epoch}.out.'$TASK_ID' \
+                -t 1-$hmm_align_njobs \
                 utils/jobarray.qsub "$cmd" $mdl_dir/split || exit 1
             find $tmpdir -name '*npy' | \
-                zip -j -@ $mdl_dir/alis.$align_no.npz > /dev/null || exit 1
-            rm -fr $tmpdir
-        else
-            echo "Using alignments: $mdl_dir/alis.$align_no.npz"
+                  zip -j -@ $mdl_dir/alis.npz > /dev/null || exit 1
         fi
 
         echo "Training the emissions"
         cmd="python -u utils/hmm-train-with-alignments.py \
                 --batch-size $hmm_train_emissions_batch_size \
-                --epochs $hmm_train_emissions_epochs \
                 --lrate $hmm_train_emissions_lrate \
                 $hmm_train_emissions_opts \
-                $mdl_dir/$mdl \
-                $mdl_dir/alis.$align_no.npz \
+                $mdl_dir/$((epoch - 1)) \
+                $mdl_dir/alis.npz \
                 $data_train_dir/feats.npz \
                 $data_train_dir/feats.stats.npz \
-                $mdl_dir/$((align_no + 1)).mdl"
+                $mdl_dir/${epoch}.mdl"
         qsub \
             -N "beer-train-emissions" \
             -cwd -j y \
-            -o $mdl_dir/log/train_emissions.$align_no.out \
+            -o $mdl_dir/log/train_emissions.${epoch}.out \
             -sync y \
             $hmm_train_emissions_sge_opts \
             utils/job.qsub "$cmd" || exit 1
 
-        # Retrieve the last model (ordered by number).
-        mdl=$(find $mdl_dir -name "[0-9]*mdl" -exec basename {} \; | \
-            sort -t '.' -k 1 -g | tail -1)
-        align_no="${mdl%.*}"
+        mdl=$mdl_dir/${epoch}.mdl
     done
+
+    cp $mdl_dir/$mdl $mdl_dir/final.mdl
 else
     echo "Model already trained: $mdl_dir/final.mdl"
 fi
